requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# 配置
pydantic-settings>=2.1.0
email-validator>=2.1.0

# 数据库
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Redis
redis>=5.0.0

# Celery
celery[redis]>=5.3.0

# 认证（模块一）
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# 工具
python-multipart>=0.0.6
Pillow
imageio
imageio-ffmpeg
